+++ src/endpoints/healthcheck.py
@@ -1,0 +1,7 @@
+from fastapi import APIRouter
+
+router = APIRouter()
+
+@router.get("/healthcheck")
+async def health_check():
+    return {"status": 200}
